+++ PrayerApp/Views/QuickAddPage.xaml.cs
@@ -2,7 +2,7 @@ using PrayerApp.ViewModels;
 
 namespace PrayerApp.Views;
 
-public partial class QuickAddPage : ContentPage
+public partial class QuickAddPage : ContentPage, IPageSheetModal
 {
     public QuickAddPage(QuickAddViewModel vm)
     {