+++ PrayerApp/Views/Tags/TagDetailPage.xaml.cs
@@ -24,8 +24,19 @@ namespace PrayerApp.Views.Tags
             BindingContext = vm;
         }
 
+        protected override void OnAppearing()
+        {
+            base.OnAppearing();
+#if IOS
+            Platforms.iOS.Helpers.SwipeBackHelper.DisableSwipeBack(this);
+#endif
+        }
+
         protected override void OnDisappearing()
         {
+#if IOS
+            Platforms.iOS.Helpers.SwipeBackHelper.EnableSwipeBack(this);
+#endif
             base.OnDisappearing();
 #if IOS
             // Detach all native gesture recognizers before the page is popped.